+++ govcd/api_vcd_test.go
@@ -167,11 +167,13 @@ type TestConfig struct {
 			SizeForUpdate int64 `yaml:"sizeForUpdate,omitempty"`
 		}
 		Nsxt struct {
-			Manager         string `yaml:"manager"`
-			Tier0router     string `yaml:"tier0router"`
-			Tier0routerVrf  string `yaml:"tier0routerVrf"`
-			Vdc             string `yaml:"vdc"`
-			ExternalNetwork string `yaml:"externalNetwork"`
+			Manager           string `yaml:"manager"`
+			Tier0router       string `yaml:"tier0router"`
+			Tier0routerVrf    string `yaml:"tier0routerVrf"`
+			Vdc               string `yaml:"vdc"`
+			ExternalNetwork   string `yaml:"externalNetwork"`
+			EdgeGateway       string `yaml:"edgeGateway"`
+			NsxtImportSegment string `yaml:"nsxtImportSegment"`
 		} `yaml:"nsxt"`
 	} `yaml:"vcd"`
 	Logging struct {
@@ -205,6 +207,7 @@ type TestVCD struct {
 	client         *VCDClient
 	org            *Org
 	vdc            *Vdc
+	nsxtVdc        *Vdc
 	vapp           *VApp
 	config         TestConfig
 	skipVappTests  bool
@@ -566,8 +569,13 @@ func (vcd *TestVCD) SetUpSuite(check *C) {
 		panic(err)
 	}
 
-	// This is set explicitly to make sure that catalog is using storage profile from correct VDC
-	setCatalogStorageProfile(vcd, check)
+	// configure NSX-T VDC for convenience if it is specified in configuration
+	if config.VCD.Nsxt.Vdc != "" {
+		vcd.nsxtVdc, err = vcd.org.GetVDCByName(config.VCD.Nsxt.Vdc, false)
+		if err != nil {
+			panic(fmt.Errorf("error geting NSX-T VDC '%s': %s", config.VCD.Nsxt.Vdc, err))
+		}
+	}
 
 	// If neither the vApp or VM tags are set, we also skip the
 	// creation of the default vApp
@@ -608,28 +616,6 @@ func (vcd *TestVCD) SetUpSuite(check *C) {
 		vcd.skipVappTests = true
 		fmt.Println("Skipping all vapp tests because one of the following wasn't given: Network, StorageProfile, Catalog, Catalogitem")
 	}
-}
-
-// setCatalogStorageProfile ensures that pre-created catalog uses storage profile from the same VDC
-// Having a storage profile from different VDC would return random errors because of not accessible storage with templates
-func setCatalogStorageProfile(vcd *TestVCD, check *C) {
-	adminOrg, err := vcd.client.GetAdminOrgByName(vcd.config.VCD.Org)
-	check.Assert(err, IsNil)
-
-	adminCatalog, err := adminOrg.GetAdminCatalogByName(vcd.config.VCD.Catalog.Name, true)
-	check.Assert(err, IsNil)
-
-	// Explicitly set catalog to use vcd.config.VCD.StorageProfile.SP1 because having multiple VDCs
-	adminCatalog.AdminCatalog.CatalogStorageProfiles = &types.CatalogStorageProfiles{[]*types.Reference{&types.Reference{
-		HREF: vcd.vdc.Vdc.VdcStorageProfiles.VdcStorageProfile[0].HREF,
-		Name: vcd.vdc.Vdc.VdcStorageProfiles.VdcStorageProfile[0].Name,
-	}}}
-
-	fmt.Printf("Setting catalog '%s' to use '%s' storage profile from VDC '%s'\n",
-		adminCatalog.AdminCatalog.Name, vcd.vdc.Vdc.VdcStorageProfiles.VdcStorageProfile[0].Name, vcd.config.VCD.Vdc)
-
-	err = adminCatalog.Update()
-	check.Assert(err, IsNil)
 }
 
 // Shows the detail of cleanup operations only if the relevant verbosity
@@ -1785,6 +1771,14 @@ func skipNoNsxtConfiguration(vcd *TestVCD, check *C) {
 		check.Skip(generalMessage + "No network pool specified")
 	}
 
+	if vcd.config.VCD.Nsxt.Vdc == "" {
+		check.Skip(generalMessage + "No NSX-T VDC specified")
+	}
+
+	if vcd.config.VCD.Nsxt.NsxtImportSegment == "" {
+		check.Skip(generalMessage + "No NSX-T Unused segment (for imported Org VDC network) specified")
+	}
+
 	if vcd.config.VCD.NsxtProviderVdc.StorageProfile == "" {
 		check.Skip(generalMessage + "No storage profile specified")
 	}
@@ -1801,6 +1795,9 @@ func skipNoNsxtConfiguration(vcd *TestVCD, check *C) {
 		check.Skip(generalMessage + "No VRF NSX-T Tier-0 router specified")
 	}
 
+	if vcd.config.VCD.Nsxt.EdgeGateway == "" {
+		check.Skip(generalMessage + "No NSX-T Edge Gateway specified in configuration")
+	}
 }
 
 // skipOpenApiEndpointTest is a helper to skip tests for particular unsupported OpenAPI endpoints