+++ govcd/vdc.go
@@ -1064,3 +1064,65 @@ func (vdc *Vdc) CreateStandaloneVMFromTemplate(params *types.InstantiateVmTempla
 	}
 	return vdc.getVmFromTask(task, params.Name)
 }
+
+// GetCapabilities allows to retrieve a list of VDC capabilities. It has a list of values. Some particularly useful are:
+// * networkProvider - overlay stack responsible for providing network functionality. (NSX_V or NSX_T)
+// * crossVdc - supports cross vDC network creation
+func (vdc *Vdc) GetCapabilities() ([]types.VdcCapability, error) {
+	if vdc.Vdc.ID == "" {
+		return nil, fmt.Errorf("VDC ID must be set to get capabilities")
+	}
+
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointVdcCapabilities
+	minimumApiVersion, err := vdc.client.checkOpenApiEndpointCompatibility(endpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	urlRef, err := vdc.client.OpenApiBuildEndpoint(fmt.Sprintf(endpoint, url.QueryEscape(vdc.Vdc.ID)))
+	if err != nil {
+		return nil, err
+	}
+
+	capabilities := make([]types.VdcCapability, 0)
+	err = vdc.client.OpenApiGetAllItems(minimumApiVersion, urlRef, nil, &capabilities)
+	if err != nil {
+		return nil, err
+	}
+	return capabilities, nil
+}
+
+// IsNsxt is a convenience function to check if VDC is backed by NSX-T pVdc
+// If error occurs - it returns false
+func (vdc *Vdc) IsNsxt() bool {
+	vdcCapabilities, err := vdc.GetCapabilities()
+	if err != nil {
+		return false
+	}
+
+	networkProviderCapability := getCapabilityValue(vdcCapabilities, "networkProvider")
+	return networkProviderCapability == types.VdcCapabilityNetworkProviderNsxt
+}
+
+// IsNsxv is a convenience function to check if VDC is backed by NSX-V pVdc
+// If error occurs - it returns false
+func (vdc *Vdc) IsNsxv() bool {
+	vdcCapabilities, err := vdc.GetCapabilities()
+	if err != nil {
+		return false
+	}
+
+	networkProviderCapability := getCapabilityValue(vdcCapabilities, "networkProvider")
+	return networkProviderCapability == types.VdcCapabilityNetworkProviderNsxv
+}
+
+// getCapabilityValue helps to lookup a specific capability in []types.VdcCapability by provided fieldName
+func getCapabilityValue(capabilities []types.VdcCapability, fieldName string) string {
+	for _, field := range capabilities {
+		if field.Name == fieldName {
+			return field.Value.(string)
+		}
+	}
+
+	return ""
+}