+++ govcd/vm_test.go
@@ -562,7 +562,7 @@ func (vcd *TestVCD) Test_VMChangeCPUCountWithCore(check *C) {
 		check.Skip("skipping test because no VM is found")
 	}
 
-	currentCpus := 0
+	currentCpus := int64(0)
 	currentCores := 0
 
 	// save current values
@@ -583,9 +583,9 @@ func (vcd *TestVCD) Test_VMChangeCPUCountWithCore(check *C) {
 	check.Assert(err, IsNil)
 
 	cores := 2
-	cpuCount := 4
+	cpuCount := int64(4)
 
-	task, err := vm.ChangeCPUCountWithCore(cpuCount, &cores)
+	task, err := vm.ChangeCPUCountWithCore(int(cpuCount), &cores)
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
@@ -607,7 +607,7 @@ func (vcd *TestVCD) Test_VMChangeCPUCountWithCore(check *C) {
 	}
 
 	// return to previous value
-	task, err = vm.ChangeCPUCountWithCore(currentCpus, &currentCores)
+	task, err = vm.ChangeCPUCountWithCore(int(currentCpus), &currentCores)
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)