+++ types/v56/nsxt_types.go
@@ -104,3 +104,91 @@ type OpenAPIEdgeGatewayEdgeClusterConfig struct {
 	PrimaryEdgeCluster   OpenAPIEdgeGatewayEdgeCluster `json:"primaryEdgeCluster,omitempty"`
 	SecondaryEdgeCluster OpenAPIEdgeGatewayEdgeCluster `json:"secondaryEdgeCluster,omitempty"`
 }
+
+// OpenApiOrgVdcNetwork allows to manage Org Vdc networks
+type OpenApiOrgVdcNetwork struct {
+	ID          string `json:"id,omitempty"`
+	Name        string `json:"name"`
+	Description string `json:"description"`
+	Status      string `json:"status,omitempty"`
+	// OwnerRef defines Org VDC or VDC Group that this network belongs to. If the ownerRef is set to a VDC Group, this
+	// network will be available across all the VDCs in the vDC Group. If the VDC Group is backed by a NSX-V network
+	// provider, the Org VDC network is automatically connected to the distributed router associated with the VDC Group
+	// and the "connection" field does not need to be set. For API version 35.0 and above, this field should be set for
+	// network creation.
+	//
+	// Note. In lower API versions (i.e. 32.0) this field is not recognized and OrgVdc should be used instead
+	OwnerRef *OpenApiReference `json:"ownerRef,omitempty"`
+
+	OrgVdc *OpenApiReference `json:"orgVdc,omitempty"`
+
+	// NetworkType describes type of Org Vdc network. ('NAT_ROUTED', 'ISOLATED')
+	NetworkType string `json:"networkType"`
+
+	// Connection specifies the edge gateway this network is connected to.
+	//
+	// Note. When NetworkType == ISOLATED, there is no uplink connection.
+	Connection *Connection `json:"connection,omitempty"`
+
+	// backingNetworkId contains the NSX ID of the backing network.
+	BackingNetworkId string `json:"backingNetworkId,omitempty"`
+	// backingNetworkType contains object type of the backing network. ('VIRTUAL_WIRE' for NSX-V, 'NSXT_FLEXIBLE_SEGMENT'
+	// for NSX-T)
+	BackingNetworkType string `json:"backingNetworkType,omitempty"`
+
+	// ParentNetwork should have external network ID specified when creating NSX-V direct network
+	ParentNetwork *OpenApiReference `json:"parentNetworkId"`
+
+	// GuestVlanTaggingAllowed specifies whether guest VLAN tagging is allowed
+	GuestVlanTaggingAllowed *bool `json:"guestVlanTaggingAllowed"`
+
+	// Subnets contains list of subnets defined on
+	Subnets OrgVdcNetworkSubnets `json:"subnets"`
+
+	// SecurityGroups defines a list of firewall groups of type SECURITY_GROUP that are assigned to the Org VDC Network.
+	// These groups can then be used in firewall rules to protect the Org VDC Network and allow/deny traffic.
+	SecurityGroups *OpenApiReferences `json:"securityGroups,omitempty"`
+
+	// RouteAdvertised reports if this network is advertised so that it can be routed out to the external networks. This
+	// applies only to network backed by NSX-T. Value will be unset if route advertisement is not applicable.
+	RouteAdvertised *bool `json:"routeAdvertised,omitempty"`
+
+	// TotalIpCount is a read only attribute reporting total number of IPs available in network
+	TotalIpCount *int `json:"totalIpCount"`
+
+	// UsedIpCount is a read only attribute reporting number of used IPs in network
+	UsedIpCount *int `json:"usedIpCount"`
+
+	// Shared shares network with other VDCs in the organization
+	Shared *bool `json:"shared,omitempty"`
+}
+
+// OrgVdcNetworkSubnetIPRanges is a type alias to reuse the same definitions with appropriate names
+type OrgVdcNetworkSubnetIPRanges = ExternalNetworkV2IPRanges
+
+// OrgVdcNetworkSubnetIPRangeValues is a type alias to reuse the same definitions with appropriate names
+type OrgVdcNetworkSubnetIPRangeValues = ExternalNetworkV2IPRange
+
+//OrgVdcNetworkSubnets
+type OrgVdcNetworkSubnets struct {
+	Values []OrgVdcNetworkSubnetValues `json:"values"`
+}
+
+type OrgVdcNetworkSubnetValues struct {
+	Gateway      string                      `json:"gateway"`
+	PrefixLength int                         `json:"prefixLength"`
+	DNSServer1   string                      `json:"dnsServer1"`
+	DNSServer2   string                      `json:"dnsServer2"`
+	DNSSuffix    string                      `json:"dnsSuffix"`
+	IPRanges     OrgVdcNetworkSubnetIPRanges `json:"ipRanges"`
+}
+
+// Connection specifies the edge gateway this network is connected to
+type Connection struct {
+	RouterRef      OpenApiReference `json:"routerRef"`
+	ConnectionType string           `json:"connectionType,omitempty"`
+}
+
+// NsxtImportableSwitch is a type alias with better name for holding NSX-T Segments (Logical Switches) which can be used
+// to back NSX-T imported Org VDC network
+type NsxtImportableSwitch = OpenApiReference